PUT: /v1/users/me GET: /v1/notification

+++ src/main/java/javapro/api/response/FileStorageResponse.java
@@ -23,6 +23,6 @@ public class FileStorageResponse {
 
     private String fileType;
 
-    private int createdAt;
+    private Long createdAt;
 
 }
